Silent migrate MQTT protocol version to version 5 if the broker supports it or raise an issue
