+++ homeassistant/components/mqtt/repairs.py
@@ -5,12 +5,10 @@ from typing import TYPE_CHECKING
 import voluptuous as vol
 
 from homeassistant.components.repairs import RepairsFlow, RepairsFlowResult
-from homeassistant.const import CONF_PORT, CONF_PROTOCOL
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import device_registry as dr
 
-from .config_flow import try_connection
-from .const import DEFAULT_PORT, DOMAIN, PROTOCOL_5
+from .const import DOMAIN
 
 URL_MQTT_BROKER_CONFIGURATION = (
     "https://www.home-assistant.io/integrations/mqtt/#broker-configuration"
@@ -55,55 +53,6 @@ class MQTTDeviceEntryMigration(RepairsFlow):
         )
 
 
-class MQTTProtocolV5Migration(RepairsFlow):
-    """Handler to migrate to MQTT protocol version 5."""
-
-    def __init__(self, entry_id: str, broker: str, protocol: str) -> None:
-        """Initialize the flow."""
-        self.entry_id = entry_id
-        self.broker = broker
-        self.protocol = protocol
-
-    async def async_step_init(
-        self, user_input: dict[str, str] | None = None
-    ) -> RepairsFlowResult:
-        """Handle the first step of a fix flow."""
-        return await self.async_step_confirm()
-
-    async def async_step_confirm(
-        self, user_input: dict[str, str] | None = None
-    ) -> RepairsFlowResult:
-        """Handle the confirm step of a fix flow."""
-        if user_input is not None:
-            entry = self.hass.config_entries.async_get_entry(self.entry_id)
-            if TYPE_CHECKING:
-                assert entry is not None
-            new_entry_data = entry.data.copy()
-            new_entry_data[CONF_PROTOCOL] = PROTOCOL_5
-            # Try the connection with protocol version 5
-            if await self.hass.async_add_executor_job(
-                try_connection,
-                {CONF_PORT: DEFAULT_PORT} | new_entry_data,
-            ):
-                self.hass.config_entries.async_update_entry(entry, data=new_entry_data)
-                return self.async_create_entry(data={})
-
-            return self.async_abort(
-                reason="mqtt_broker_migration_to_v5_failed",
-                description_placeholders={
-                    "broker": self.broker,
-                    "protocol": self.protocol,
-                    "url_mqtt_broker_configuration": URL_MQTT_BROKER_CONFIGURATION,
-                },
-            )
-
-        return self.async_show_form(
-            step_id="confirm",
-            data_schema=vol.Schema({}),
-            description_placeholders={"broker": self.broker, "protocol": self.protocol},
-        )
-
-
 async def async_create_fix_flow(
     hass: HomeAssistant,
     issue_id: str,
@@ -113,10 +62,6 @@ async def async_create_fix_flow(
     if TYPE_CHECKING:
         assert data is not None
     entry_id: str = data["entry_id"]  # type: ignore[assignment]
-    if issue_id == "protocol_5_migration":
-        broker: str = data["broker"]  # type: ignore[assignment]
-        protocol: str = data["protocol"]  # type: ignore[assignment]
-        return MQTTProtocolV5Migration(entry_id, broker, protocol)
     subentry_id: str = data["subentry_id"]  # type: ignore[assignment]
     name: str = data["name"]  # type: ignore[assignment]
     return MQTTDeviceEntryMigration(
